Remove TooGenericExceptionSpec
