+++ detekt-rules-exceptions/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/exceptions/TooGenericExceptionThrownSpec.kt
@@ -87,4 +87,32 @@ class TooGenericExceptionThrownSpec {
 
         assertThat(rule.compileAndLint(code)).isEmpty()
     }
+
+    @Test
+    fun `should not report any`() {
+        val rule = TooGenericExceptionThrown(TestConfig(EXCEPTION_NAMES to "[]"))
+        val code = """
+            fun main() {
+                try {
+                    throw Throwable()
+                } catch (e: ArrayIndexOutOfBoundsException) {
+                    throw Error()
+                } catch (e: Error) {
+                    throw Exception()
+                } catch (e: Exception) {
+                } catch (e: IllegalMonitorStateException) {
+                } catch (e: IndexOutOfBoundsException) {
+                    throw RuntimeException()
+                } catch (e: Throwable) {
+                } catch (e: RuntimeException) {
+                    throw NullPointerException()
+                } catch (e: NullPointerException) {
+
+                }
+            }
+        """.trimIndent()
+        val findings = rule.compileAndLint(code)
+
+        assertThat(findings).isEmpty()
+    }
 }
