--- detekt-rules-exceptions/src/test/resources/deactivated-exceptions.yml
@@ -1,10 +1,0 @@
-exceptions:
-  active: true
-  TooGenericExceptionCaught:
-    active: true
-    exceptionNames:
-      -
-  TooGenericExceptionThrown:
-    active: true
-    exceptionNames:
-      -
